Extract device-page tabs into components (Files/Tweaks/Sideload/Remote/Optimize)
Pure move: the 3,950-line device page drops to 2,379 by peeling the five
self-contained tabs into $lib/components, ~2,050 lines total. Each
component takes `serial`, loads its own data in onMount, and mounts once
per device visit (parent keeps a `visited` map and toggles `hidden`), so
state still survives tab switches and an empty result still means
"loaded" — the lazy-load $effect now only covers the in-page tabs.

OptimizeTab keeps cross-tab parity through three explicit seams:
`resetToken` (parent bumps it where it used to null the plan),
`onStatesChanged` (re-fetch appStates after a run), and `onPlanLoaded`
(lazy usage cues). App-List-only recommendation helpers stay in the page.

In-scope fix: RemoteTab clears its typing flush timer in onDestroy (the
page-level cleanup always missed it). package-lock version sync rides
along (it still said 0.1.0-beta.8).
